Fix: Subsonic: Handle uninitialized port config value (#2263)
Fix uninitialized port config value

The or construction does not seem to protect against None value when
there are tranformations applied in the lhs (or evaluates after the
str() and int() conversions. This leads to problems when no value is set
for port. Change back to the more verbose if port else 443 construction
to fix this.

Fixes: https://github.com/music-assistant/support/issues/4086

Signed-off-by: Eric B Munson <eric@munsonfam.org>