Add codec setting for Sendspin over Cast (#2790)
Add a option to select between flac, pcm, and opus.

Flac works from my limited testing on 3 devices, so I set it as the default now.

Opus doesn't work natively, using that option tries to decode it in software. But from my testing this is too CPU intensive for a Google Home mini.